Backdoor: Call the correct function.
With the new way how AddCodePatch is processed in ReassemblerBackend,
the sha1 function is no longer flattened into one function that starts
at the beginning of the inserted block. Calling `sha1_block` is in fact
calling the ROTATE_LEFT function. Fixing the call to `sha1` solves the
failed test case.